src/kuzu/requirements.txt
# Core dependencies
kuzu==0.11.3
polars>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0

# BAML for structured LLM interactions
baml-py==0.203.1

# LLM providers (optional, for pipeline mode)
google-genai>=1.0.0
google-generativeai>=0.5.0
anthropic>=0.25.0
